requirements.txt
# Core
python-dotenv
pydantic
pyyaml

# GenAI / LLMs
anthropic
langchain
langgraph

# Utilities
tqdm
loguru

# Testing
pytest

# APIs 
requests